+++ CodeStarterCamp_Week5/BadPersonality.swift
@@ -1,0 +1,12 @@
+//
+//  BadPersonality.swift
+//  Week5 Step1
+//
+//  Created by JeonSangHyeok on 2023/03/22.
+//
+
+import Foundation
+
+protocol BadPersonality {
+    var frequancyOfCursing: Level { get set}
+}
