+++ CodeStarterCamp_Week5/Level.swift
@@ -1,0 +1,14 @@
+//
+//  Level.swift
+//  Week5 Step1
+//
+//  Created by JeonSangHyeok on 2023/03/22.
+//
+
+import Foundation
+
+enum Level {
+    case A
+    case B
+    case C
+}
