+++ CodeStarterCamp_Week5/Person.swift
@@ -1,0 +1,18 @@
+//
+//  Person.swift
+//  Week5 Step1
+//
+//  Created by JeonSangHyeok on 2023/03/22.
+//
+
+import Foundation
+
+class Person {
+    var name: String
+    var height: Int
+    
+    init(name: String, height: Int) {
+        self.name = name
+        self.height = height
+    }
+}
