+++ CodeStarterCamp_Week5/Talent.swift
@@ -1,0 +1,14 @@
+//
+//  Talent.swift
+//  Week5 Step1
+//
+//  Created by JeonSangHyeok on 2023/03/22.
+//
+
+import Foundation
+
+protocol Talent {
+    var singing: Level { get set}
+    var dancing: Level { get set}
+    var acting: Level { get set}
+}
