+++ CodeStarterCamp_Week5/TalentedPersonWithBadPersonality.swift
@@ -1,0 +1,24 @@
+//
+//  TalentedPersonWithBadPersonality.swift
+//  Week5 Step1
+//
+//  Created by JeonSangHyeok on 2023/03/22.
+//
+
+import Foundation
+
+class TalentedPersonWithBadPersonality: Person, Talent, BadPersonality {
+    var singing: Level
+    var dancing: Level
+    var acting: Level
+    var frequancyOfCursing: Level
+    
+    init(name: String, height: Int, singing: Level, dancing: Level, acting: Level, frequancyOfCursing: Level) {
+        self.singing = singing
+        self.dancing = dancing
+        self.acting = acting
+        self.frequancyOfCursing = frequancyOfCursing
+        
+        super.init(name: name, height: height)
+    }
+}
